+++ app/Http/Controllers/UserController.php
@@ -7,5 +7,10 @@ use App\Models\User;
 
 class UserController extends Controller
 {
-    //
+    public function index(Request $request, User $user)
+    {
+        $userId = $request->input('user_id');
+        $singleUser = $user->find($userId);
+        return view('users/index')->with(['user' => $user->first()]);
+    }
 }
